v5.0-fix-zip: VersionTesting added
