+++ sigmastate/src/test/scala/sigmastate/ScriptVersionSwitchSpecification.scala
@@ -30,7 +30,7 @@ class ScriptVersionSwitchSpecification extends SigmaDslTesting {
 
   implicit def IR = createIR()
 
-  val b1 = CostingBox(
+  lazy val b1 = CostingBox(
     false,
     new ErgoBox(
       1L,
